requirements.txt
img2pdf==0.3.4
natsort==7.0.1
pdf2image==1.12.1
Pillow==7.1.1
PyPDF2==1.26.0
